+++ src/features/api/userData.types.ts
@@ -22,3 +22,27 @@ export interface IUserProfile {
 	last_login?: Date;
 	last_action?: Date;
 }
+
+export interface IUserProfilePatch {
+	fio_apikey: string | null;
+	prun_username: string | null;
+	email: string | null;
+}
+
+export interface IUserChangePasswordPayload {
+	old: string;
+	new: string;
+}
+
+export interface IUserChangePasswordResponse {
+	message: string;
+}
+
+export interface IUserVerifyEmailPayload {
+	code: string;
+}
+
+export interface IUserVerifyEmailResponse {
+	status_code: number;
+	message: string;
+}