+++ src/ui/components/PFormSeperator.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+	<div class="col-span-2">
+		<slot />
+	</div>
+</template>